feat: bill 공동 바텀모달 컴포넌트
